requirements.txt
fastapi
uvicorn
pydantic
diffusers
transformers
torch
Pillow